+++ test/mut/mutator/atom_literal_test.exs
@@ -30,11 +30,11 @@ defmodule Mut.Mutator.AtomLiteralTest do
       assert AtomLiteral.applicable?({:__block__, [], [:lt]}, ctx([]))
     end
 
-    test "false for non-allowlisted atoms, booleans, nil, and wrong context" do
+    test "false for non-allowlisted atoms/booleans/nil/match; true under schema engine (M52)" do
       refute AtomLiteral.applicable?({:__block__, [], [:something_else]}, ctx([]))
       refute AtomLiteral.applicable?({:__block__, [], [true]}, ctx([]))
       refute AtomLiteral.applicable?({:__block__, [], [nil]}, ctx([]))
-      refute AtomLiteral.applicable?({:__block__, [], [:ok]}, ctx(engine: :schema))
+      assert AtomLiteral.applicable?({:__block__, [], [:ok]}, ctx(engine: :schema))
       refute AtomLiteral.applicable?({:__block__, [], [:ok]}, ctx(env_context: :match))
     end
   end